requirements.txt
jsoneng==1.0.5
python-dotenv==1.0.0
numpy
matplotlib
termcolor
